install.sh

+++ install.sh
@@ -1,16 +1,27 @@
 apt-get install exuberant-ctags
 apt-get install tmux
 
+
 git submodule init
 git submodule update
 
-rm ~/.vimrc
+#typescript config
+cd  ~/.vim/bundle/vimproc.vim 
+make
+
+
+if [ -f ~/.vimrc ];
+then
+    rm ~/.vimrc
+fi
 ln -s ~/.vim/.vimrc ~/.vimrc
 
-rm ~/.tmux.conf
+if [ -f ~/.tmux.conf ];
+then
+    rm ~/.tmux.conf
+fi
 ln -s ~/.vim/.tmux.conf ~/.tmux.conf
 
-echo 'alias tmux="TERM=screen-256color-bce tmux"' >> ~/.bashrc &&  exec bash
+echo 'alias tmux="TERM=screen-256color-bce tmux"' >> ~/.bashrc
+exec bash
 
-#typescript config
-cd  ~/.vim/bundle/vimproc.vim &&  make
